+++ nirum/datastructures.py
@@ -6,7 +6,7 @@
 """
 import collections
 
-__all__ = 'Map',
+__all__ = 'Map', 'ImmutableList'
 
 
 class Map(collections.Mapping):
@@ -49,3 +49,27 @@ class Map(collections.Mapping):
         else:
             args = ''
         return '{0.__module__}.{0.__name__}({1})'.format(type(self), args)
+
+
+class ImmutableList(collections.Sequence):
+
+    def __init__(self, l):
+        self.l = l
+
+    def __getitem__(self, index):
+        return self.l[index]
+
+    def __len__(self):
+        return len(self.l)
+
+    def __contains__(self, item):
+        return item in self.l
+
+    def __iter__(self):
+        return iter(self.l)
+
+    def index(self, item):
+        return self.l.index(item)
+
+    def count(self):
+        return self.l.count()
